+++ app/__generated__/$resources.sitemap.xml.ts
@@ -2,7 +2,7 @@
       export const sitemap = [
   {
     "path": "/",
-    "lastModified": "2025-05-21"
+    "lastModified": "2025-05-24"
   }
 ];