+++ app/__generated__/[2022].[08].[how-to-install-home-assistant-on-virtualbox.html]._index.ts
@@ -1,0 +1,4 @@
+
+        export const url = "https://old.taikun-life.com/2022/08/how-to-install-home-assistant-on-virtualbox.html";
+        export const status = 301;
+