+++ app/__generated__/[sitemap.xml]._index.tsx
@@ -9,7 +9,7 @@
 
       export const projectId = "dba9d753-ddf9-434f-be13-fdc5fabadb2d";
 
-      export const lastPublished = "2025-05-21T12:31:36.613Z";
+      export const lastPublished = "2025-05-24T10:24:51.426Z";
 
       export const siteName = "たいくんの部屋";