+++ app/routes/[2021].[11].[used-mac-how-to-choice-guide-2022.html]._index.ts
@@ -1,0 +1,6 @@
+import { redirect } from "react-router";
+import { url, status } from "../__generated__/[2021].[11].[used-mac-how-to-choice-guide-2022.html]._index";
+
+export const loader = () => {
+  throw redirect(url, status);
+};